Use the new values internally. Switch to using current constants for internal code
