+++ docs/docs/view-config-json.md
@@ -49,6 +49,8 @@ The view config schema version.
 | `1.0.7`| The `geneAlias` option for the `anndata-expression-matrix.zarr` file type was added in this version. |
 | `1.0.8`| Support for multi-dataset views and dataset-specific coordination scope mappings was added in this version. |
 | `1.0.9`| Support for plugin coordination types was added in this version. |
+// TODO
+| `1.0.11`| Spatial layer coordination type names were changed in this version. |
 
 ### `initStrategy`
 - Type: `string`
