+++ src/app/state/coordination.js
@@ -29,10 +29,10 @@ export const DEFAULT_COORDINATION_VALUES = {
   [CoordinationType.SPATIAL_AXIS_FIXED]: false,
   [CoordinationType.SPATIAL_ROTATION_ORBIT]: 0,
   [CoordinationType.SPATIAL_ORBIT_AXIS]: 'Y',
-  [CoordinationType.SPATIAL_RASTER_LAYERS]: null,
-  [CoordinationType.SPATIAL_CELLS_LAYER]: null,
-  [CoordinationType.SPATIAL_MOLECULES_LAYER]: null,
-  [CoordinationType.SPATIAL_NEIGHBORHOODS_LAYER]: null,
+  [CoordinationType.SPATIAL_RASTER_LAYER]: null,
+  [CoordinationType.SPATIAL_SEGMENTATION_LAYER]: null,
+  [CoordinationType.SPATIAL_POINT_LAYER]: null,
+  [CoordinationType.SPATIAL_NEIGHBORHOOD_LAYER]: null,
   [CoordinationType.HEATMAP_ZOOM_X]: 0,
   [CoordinationType.HEATMAP_ZOOM_Y]: 0,
   [CoordinationType.HEATMAP_TARGET_X]: 0,
@@ -119,10 +119,10 @@ export const COMPONENT_COORDINATION_TYPES = {
     CoordinationType.DATASET,
     CoordinationType.SPATIAL_ZOOM,
     CoordinationType.SPATIAL_ROTATION,
-    CoordinationType.SPATIAL_RASTER_LAYERS,
-    CoordinationType.SPATIAL_CELLS_LAYER,
-    CoordinationType.SPATIAL_MOLECULES_LAYER,
-    CoordinationType.SPATIAL_NEIGHBORHOODS_LAYER,
+    CoordinationType.SPATIAL_RASTER_LAYER,
+    CoordinationType.SPATIAL_SEGMENTATION_LAYER,
+    CoordinationType.SPATIAL_POINT_LAYER,
+    CoordinationType.SPATIAL_NEIGHBORHOOD_LAYER,
     CoordinationType.SPATIAL_TARGET_X,
     CoordinationType.SPATIAL_TARGET_Y,
     CoordinationType.SPATIAL_TARGET_Z,
@@ -210,10 +210,10 @@ export const COMPONENT_COORDINATION_TYPES = {
   ],
   [Component.LAYER_CONTROLLER]: [
     CoordinationType.DATASET,
-    CoordinationType.SPATIAL_RASTER_LAYERS,
-    CoordinationType.SPATIAL_CELLS_LAYER,
-    CoordinationType.SPATIAL_MOLECULES_LAYER,
-    CoordinationType.SPATIAL_NEIGHBORHOODS_LAYER,
+    CoordinationType.SPATIAL_RASTER_LAYER,
+    CoordinationType.SPATIAL_SEGMENTATION_LAYER,
+    CoordinationType.SPATIAL_POINT_LAYER,
+    CoordinationType.SPATIAL_NEIGHBORHOOD_LAYER,
     CoordinationType.SPATIAL_ZOOM,
     CoordinationType.SPATIAL_TARGET_X,
     CoordinationType.SPATIAL_TARGET_Y,
@@ -240,7 +240,7 @@ export const COMPONENT_COORDINATION_TYPES = {
   ],
   [Component.DESCRIPTION]: [
     CoordinationType.DATASET,
-    CoordinationType.SPATIAL_RASTER_LAYERS,
+    CoordinationType.SPATIAL_RASTER_LAYER,
   ],
   higlass: [
     CoordinationType.DATASET,
